+++ CHANGELOG.md
@@ -7,6 +7,8 @@ FEATURES:
 * greenplum: support for changing the user's password
 * greenplum: support for changing segment and master resources
 * mdb: add `template_db` attribute to `yandex_mdb_postgresql_cluster` resource and data source
+* clickhouse: add `assign_public_ip` attribute to `host` declaration in resource and data source
+* clickhouse: support hosts update
 
 ## 0.77.0 (July 27, 2022)
 BUG FIXES: